backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic[email]
python-dotenv
python-jose[cryptography]
requests
PyJWT
alembic
cachetools